+++ src/constants/variables.ts
@@ -7,6 +7,8 @@ export const VARIABLES = {
   issueFooter: "issue-footer",
   addProjectLabel: "add-project-label",
   addComment: "add-comment",
+  closeOnSuccess: "close-on-success",
+  closeOnSuccessMsg: "close-on-success-msg",
   jobSummary: "job-summary",
   quite: "quite",
 };