+++ src/models/index.ts
@@ -1,4 +1,4 @@
-export * from "./FailedTestInfo";
 export * from "./JSONReport";
 export * from "./TestError";
+export * from "./TestInfo";
 export * from "./TestStatus";