+++ src/utils/createIssueTitle.ts
@@ -1,0 +1,10 @@
+export const createIssueTitle = (
+  suiteTilte: string,
+  specTitle: string,
+  projectName: string,
+  issuePrefix?: string
+) => {
+  return `${
+    issuePrefix ? `${issuePrefix} ` : ""
+  }${suiteTilte} - ${specTitle} (${projectName})`;
+};