+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0]
+
+- Added the `close-on-success` and `close-on-success-msg` inputs to close the issue when the test is successful.
+- Changed the assignees logic
+
 ## [1.1.0]
 
 - Added the `issue-assignees` input to assign the issue to specific users.